requirements.txt
flask
numpy
pillow
scikit-image
torch
fastai
gdown
google-cloud-storage
pytest
